commit: arregle lo que hicieron del menu, agregue la parte de añadir gato, agregue el enum vacunas y la opcion de elegir que vacuna tiene la mascota

+++ src/logica/Perro.java
@@ -9,7 +9,7 @@ public class Perro extends Canino {
     
     public Perro(String nombre, String raza, 
     byte edad, String dientes, String aullan,
-     vacunas vacunas, int costos, String origen){
+     vacunas vacunas, Double costos, String origen){
         super(dientes, aullan, vacunas, costos, origen);
         this.nombre = nombre;
         this.raza = raza;
